pull archive extraction into a separate package
